refactor: rename helper file for improving readability
