+++ test/__tests__/__integration_util_test__/v0_utilities.test.js
@@ -1,6 +1,6 @@
 
 const utilities = require("../../../src/v0/util");
-const { getFuncTestData } = require("./helper");
+const { getFuncTestData } = require("./testHelper");
 
 const functionsToTest = [
   "flattenJson",
